LayoutTests/ietestcenter/Javascript/15.12.2-0-2-expected.txt

15.12.2-0-2 (TestCases/chapter15/15.12/15.12.2/15.12.2-0-2.js) JSON.parse must exist as a function taking 2 parameters

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ES5Harness.preconditionPassed is true
FAIL ES5Harness.testPassed should be true (of type boolean). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
